refactor(blorktools): model-texture-util
